Made the logic of the method RequestMerch in  the controller

+++ src/OzonEdu.MerchandiseService.ApplicationServices/Commands/OrderAggregate/GiveOutOrderCommand.cs
@@ -4,11 +4,11 @@ namespace OzonEdu.MerchandiseService.ApplicationServices.Commands
 {
     public class GiveOutOrderCommand : IRequest<bool>
     {
-        public GiveOutOrderCommand(int orderId)
+        public GiveOutOrderCommand(long orderId)
         {
             OrderId = orderId;
         }
 
-        public int OrderId { get; private set; }
+        public long OrderId { get; private set; }
     }
 }
